[Feature][JsonSplit] modify updateProcessDefinition/switchProcessDefinitionVersion  (#4666)
* processdefinition create/delete method

* init

* add relation parse

* delete process_definition_json

* modify updateProcessDefinition/switchProcessDefinitionVersion

* codestyle

* codestyle

* fix compile error （version type）

* fix compile error （version type）

+++ dolphinscheduler-dao/src/test/java/org/apache/dolphinscheduler/dao/mapper/ProcessDefinitionMapperTest.java
@@ -358,7 +358,7 @@ public class ProcessDefinitionMapperTest {
 
     @Test
     public void testUpdateVersionByProcessDefinitionId() {
-        long expectedVersion = 10;
+        int expectedVersion = 10;
         ProcessDefinition processDefinition = insertOne();
         processDefinition.setVersion(expectedVersion);
         processDefinitionMapper.updateVersionByProcessDefinitionId(
